Organize location folder
