+++ event-details-boundary/JoinEvent.test.tsx
@@ -17,7 +17,7 @@ import { mockLocationCoordinate2D, mockTiFLocation } from "@location/MockData"
 import "@test-helpers/Matchers"
 import { TrueRegionMonitor } from "@arrival-tracking/region-monitoring/MockRegionMonitors"
 import { resetTestSQLiteBeforeEach, testSQLite } from "@test-helpers/SQLite"
-import { SQLiteRecentLocationsStorage } from "@lib/RecentsLocations"
+import { SQLiteRecentLocationsStorage } from "@location/Recents"
 import { mswServer } from "@test-helpers/msw"
 import { http, HttpResponse } from "msw"
 import { TiFAPI } from "TiFShared/api"
