+++ event-details-boundary/JoinEvent.tsx
@@ -22,7 +22,7 @@ import { useSharedValue } from "react-native-reanimated"
 import { FontScaleFactors } from "@lib/Fonts"
 import { BottomSheetBackdrop, BottomSheetModal } from "@gorhom/bottom-sheet"
 import { TiFAPI } from "TiFShared/api"
-import { RecentLocationsStorage } from "@lib/RecentsLocations"
+import { RecentLocationsStorage } from "@location/Recents"
 import { updateEventDetailsQueryEvent } from "./Query"
 import { JoinEventResponse } from "TiFShared/api/models/Event"
 import { EventLocation } from "TiFShared/domain-models/Event"
