+++ event-details-boundary/TravelEstimates.test.tsx
@@ -10,7 +10,7 @@ import {
   mockLocationCoordinate2D
 } from "@location/MockData"
 import { UserLocationFunctionsProvider } from "@location/UserLocation"
-import { EXPO_LOCATION_ERRORS } from "@shared-models/Location"
+import { EXPO_LOCATION_ERRORS } from "@location/Expo"
 import { setPlatform } from "@test-helpers/Platform"
 import { CodedError } from "expo-modules-core"
 import { LocationCoordinate2D } from "TiFShared/domain-models/LocationCoordinate2D"
