+++ location/Expo.ts
@@ -1,0 +1,4 @@
+export const EXPO_LOCATION_ERRORS = {
+  noPermissions: "E_NO_PERMISSIONS",
+  servicesDisabled: "E_LOCATION_SERVICES_DISABLED"
+}
