+++ location/Geocoding.tsx
@@ -1,5 +1,5 @@
 import React, { ReactNode, createContext, useContext } from "react"
-import { TiFLocation } from "."
+import { NamedLocation } from "./NamedLocation"
 import { useQuery } from "@tanstack/react-query"
 import { QueryHookOptions } from "@lib/ReactQuery"
 import { LocationCoordinate2D } from "TiFShared/domain-models/LocationCoordinate2D"
@@ -9,7 +9,7 @@ import { LocationCoordinate2D } from "TiFShared/domain-models/LocationCoordinate
  */
 export const useReverseGeocodeQuery = (
   coordinates: LocationCoordinate2D,
-  options: QueryHookOptions<TiFLocation> = DEFAULT_REVERSE_GEOCODE_QUERY_OPTIONS
+  options: QueryHookOptions<NamedLocation> = DEFAULT_REVERSE_GEOCODE_QUERY_OPTIONS
 ) => {
   const { reverseGeocode } = useGeocodingFunctions()
   return useQuery(
@@ -32,7 +32,7 @@ export const DEFAULT_REVERSE_GEOCODE_QUERY_OPTIONS = {
  * A set of functions required for performing geocoding operations.
  */
 export type GeocodingFunctions = {
-  reverseGeocode: (coordinates: LocationCoordinate2D) => Promise<TiFLocation>
+  reverseGeocode: (coordinates: LocationCoordinate2D) => Promise<NamedLocation>
 }
 
 const GeocodingFunctionsContext = createContext<GeocodingFunctions | undefined>(
