+++ location/index.ts
@@ -1,3 +1,2 @@
-export * from "@shared-models/Location"
 export * from "./Geocoding"
 export * from "./UserLocation"
